:recycle:Refactor missions code
